Handle the location permission properly
We can’t start monitoring without a location
permission, so we have to wait until the permission
is granted by the user.

It is expected that the calling app will request
a permission from the user in the app’s UI on Android 6+
devices.

Note that the permission can also be revoked or granted
back by the user through the system settings.
We will start monitoring in all those cases as soon as the app
is brought to foreground or restarted by the user.