Fix a potential crash on shutdown
This would happen in the following edge case:
* No location permission has been granted by the user
* mParticle config changed to disable Skyhook or change the key

The exception thrown by the Skyhook Kit in this case
would have been propagated straight to the calling app
because mParticle SDK doesn't handle exceptions
from the onKitDestroy() call.

In addition to the above, this commit handles other edge
cases where onKitDestroy() or onActivityResumed() are
called after onKitCreate() threw an exception.